Planets and Background
Made the background into space, this is for now only the home page but will become all pages' backgrounds. This animates by floating to the left. There are 5 planets on this page each of which will become a page in the future. Each planet is different but will be made to rotate. This page is the homepage, more work needs to be done, my next step will be adding a title to each planet.