Services
Made the base layout for the services page. This simple layout took as long as it did because of issues I was having when it came t using Bootstrap's carousel. Text for the projects still needed to be added as well as image representations. After those then links to the GitHub will be put in.