Resume Page
Started the resume page, it's almost completely done. The only thing it needs is a bit more to make it unique. One thing that has come to mind for all my pages is my space theme. I need to find a way for each page to have a unique version of the theme.